Change migration for heroku

+++ db/migrate/20141224143938_remove_created_by_index_from_tickets.rb
@@ -1,7 +1,7 @@
 class RemoveCreatedByIndexFromTickets < ActiveRecord::Migration
   def change
     change_table :tickets do |t|
-      t.remove_index :created_at_and_created_by
+      t.remove_index :created_at_and_created_by unless Rails.env == "production"
       t.index :created_at, unique: true
       t.index [:created_at, :user_id], unique: true
     end
